// src/app/api/generated/models/employees-dto.ts
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

export interface EmployeesDto {
  address?: (string | null);
  birthDate?: (string | null);
  city?: (string | null);
  country?: (string | null);
  employeeID: (number | string);
  extension?: (string | null);
  firstName: (string | null);
  hireDate?: (string | null);
  homePhone?: (string | null);
  lastName: (string | null);
  notes?: (string | null);
  photo?: (string | null);
  photoPath?: (string | null);
  postalCode?: (string | null);
  region?: (string | null);
  reportsTo?: (number | string | null);
  title?: (string | null);
  titleOfCourtesy?: (string | null);
}